requirements.txt
flask
openpyxl
python-barcode
Pillow
